Fixed a bug with previously deferred tasks.
Previously the script was hiding tasks based on the existence of a
defer date. This meant that a task that was once deferred but is now
available would not show up in the widget.

This commit alters the script to compare against today and only hide
deferred tasks if they truly are still deferred.